backend/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
psycopg2-binary
alembic
pydantic
openai
google-generativeai
python-multipart
python-dotenv
httpx
